+++ pybop/parameters/priors.py
@@ -56,6 +56,38 @@ class BasePrior:
         """
         return self.prior.logpdf(x, loc=self.loc, scale=self.scale)
 
+    def icdf(self, q):
+        """
+        Calculates the inverse cumulative distribution function (CDF) of the distribution at q.
+
+        Parameters
+        ----------
+        q : float
+            The point(s) at which to evaluate the inverse CDF.
+
+        Returns
+        -------
+        float
+            The inverse cumulative distribution function value at q.
+        """
+        return self.prior.ppf(q, scale=self.scale, loc=self.loc)
+
+    def cdf(self, x):
+        """
+        Calculates the cumulative distribution function (CDF) of the distribution at x.
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the CDF.
+
+        Returns
+        -------
+        float
+            The cumulative distribution function value at x.
+        """
+        return self.prior.cdf(x, scale=self.scale, loc=self.loc)
+
     def rvs(self, size=1, random_state=None):
         """
         Generates random variates from the distribution.
@@ -120,6 +152,10 @@ class BasePrior:
         """
         return self.scale
 
+    @property
+    def n_parameters(self):
+        return self._n_parameters
+
 
 class Gaussian(BasePrior):
     """
@@ -141,6 +177,44 @@ class Gaussian(BasePrior):
         self.loc = mean
         self.scale = sigma
         self.prior = stats.norm
+        self._offset = -0.5 * np.log(2 * np.pi * self.scale**2)
+        self.sigma2 = self.scale**2
+        self._multip = -1 / (2.0 * self.sigma2)
+        self._n_parameters = 1
+
+    def __call__(self, x):
+        """
+        Evaluates the gaussian (log) distribution at x.
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the distribution.
+
+        Returns
+        -------
+        float
+            The value(s) of the distribution at x.
+        """
+        return self.logpdf(x)
+
+    def evaluateS1(self, x):
+        """
+        Evaluates the first derivative of the gaussian (log) distribution at x.
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the first derivative.
+
+        Returns
+        -------
+        float
+            The value(s) of the first derivative at x.
+        """
+        if not isinstance(x, np.ndarray):
+            x = np.asarray(x)
+        return self(x), -(x - self.loc) * self._multip
 
 
 class Uniform(BasePrior):
@@ -165,6 +239,41 @@ class Uniform(BasePrior):
         self.loc = lower
         self.scale = upper - lower
         self.prior = stats.uniform
+        self._n_parameters = 1
+
+    def __call__(self, x):
+        """
+        Evaluates the gaussian (log) distribution at x.
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the distribution.
+
+        Returns
+        -------
+        float
+            The value(s) of the distribution at x.
+        """
+        return self.logpdf(x)
+
+    def evaluateS1(self, x):
+        """
+        Evaluates the first derivative of the log uniform distribution at x.
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the first derivative.
+
+        Returns
+        -------
+        float
+            The value(s) of the first derivative at x.
+        """
+        log_pdf = self.__call__(x)
+        dlog_pdf = np.zeros_like(x)
+        return log_pdf, dlog_pdf
 
     @property
     def mean(self):
@@ -199,3 +308,105 @@ class Exponential(BasePrior):
         self.loc = loc
         self.scale = scale
         self.prior = stats.expon
+        self._n_parameters = 1
+
+    def __call__(self, x):
+        """
+        Evaluates the gaussian (log) distribution at x.
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the distribution.
+
+        Returns
+        -------
+        float
+            The value(s) of the distribution at x.
+        """
+        return self.logpdf(x)
+
+    def evaluateS1(self, x):
+        """
+        Evaluates the first derivative of the log exponential distribution at x.
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the first derivative.
+
+        Returns
+        -------
+        float
+            The value(s) of the first derivative at x.
+        """
+        log_pdf = self.__call__(x)
+        dlog_pdf = -1 / self.scale * np.ones_like(x)
+        return log_pdf, dlog_pdf
+
+
+class ComposedLogPrior(BasePrior):
+    """
+    Represents a composition of multiple prior distributions.
+    """
+
+    def __init__(self, *priors):
+        self._priors = priors
+        for prior in priors:
+            if not isinstance(prior, BasePrior):
+                raise ValueError("All priors must be instances of BasePrior")
+
+        self._n_parameters = len(priors)  # Needs to be updated
+
+    def __call__(self, x):
+        """
+        Evaluates the composed prior distribution at x.
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the distribution.
+
+        Returns
+        -------
+        float
+            The value(s) of the distribution at x.
+        """
+        if not isinstance(x, np.ndarray):
+            x = np.asarray(x)
+        return sum(prior(x) for prior, x in zip(self._priors, x))
+
+    def evaluateS1(self, x):
+        """
+        Evaluates the first derivative of the composed prior distribution at x.
+        Inspired by PINTS implementation.
+
+        *This method only works if the underlying :class:`LogPrior` classes all
+        implement the optional method :class:`LogPDF.evaluateS1().`.*
+
+        Parameters
+        ----------
+        x : float
+            The point(s) at which to evaluate the first derivative.
+
+        Returns
+        -------
+        float
+            The value(s) of the first derivative at x.
+        """
+        output = 0
+        doutput = np.zeros(self.n_parameters)
+        index = 0
+
+        for prior in self._priors:
+            num_params = prior.n_parameters
+            x_subset = x[index : index + num_params]
+            p, dp = prior.evaluateS1(x_subset)
+            output += p
+            doutput[index : index + num_params] = dp
+            index += num_params
+
+        return output, doutput
+
+    def __repr__(self):
+        return f"{self.__class__.__name__}, priors: {self._priors}"